[Issue-1558] Update UI for RestoreJson and ExportAll feature
