+++ src/screens/Account/AccountsScreen/index.tsx
@@ -37,7 +37,6 @@ import { isEthereumAddress } from '@polkadot/util-crypto';
 import DeleteModal from 'components/common/Modal/DeleteModal';
 import useConfirmModal from 'hooks/modal/useConfirmModal';
 import useGoHome from 'hooks/screen/useGoHome';
-import { AccountExportSelector } from 'screens/Account/AccountExportSelector';
 
 const renderListEmptyComponent = () => {
   return (
@@ -67,7 +66,6 @@ export const AccountsScreen = ({
     params: { pathName },
   },
 }: AccountsScreenProps) => {
-  const accountExportRef = useRef<ModalRef>();
   const toast = useToast();
   const theme = useSubWalletTheme().swThemes;
   const goHome = useGoHome();
@@ -337,7 +335,7 @@ export const AccountsScreen = ({
         placeholder={i18n.placeholder.accountName}
         rightIconOption={{
           icon: ({ color }) => <Icon phosphorIcon={Export} weight={'fill'} iconColor={color} size={'md'} />,
-          onPress: () => accountExportRef && accountExportRef.current?.onOpenModal(),
+          onPress: () => navigation.navigate('ExportAllAccount'),
         }}
       />
 
@@ -425,8 +423,6 @@ export const AccountsScreen = ({
         onCompleteModal={onCompleteDeleteModal}
         setVisible={setDeleteVisible}
       />
-
-      <AccountExportSelector items={accounts} accountExportRef={accountExportRef} />
     </>
   );
 };
